+++ ContentPatcher/Framework/Tokens/ValueProviders/DynamicTokenValueProvider.cs
@@ -1,5 +1,6 @@
 using System;
 using System.Collections.Generic;
+using System.Linq;
 using Pathoschild.Stardew.Common.Utilities;
 
 namespace ContentPatcher.Framework.Tokens.ValueProviders
@@ -11,7 +12,7 @@ namespace ContentPatcher.Framework.Tokens.ValueProviders
         ** Fields
         *********/
         /// <summary>The allowed root values (or <c>null</c> if any value is allowed).</summary>
-        private readonly InvariantHashSet AllowedRootValues;
+        private InvariantHashSet AllowedRootValues;
 
         /// <summary>The current values.</summary>
         private InvariantHashSet Values = new InvariantHashSet();
@@ -31,18 +32,28 @@ namespace ContentPatcher.Framework.Tokens.ValueProviders
 
         /// <summary>Add a set of possible values.</summary>
         /// <param name="possibleValues">The possible values to add.</param>
-        public void AddAllowedValues(InvariantHashSet possibleValues)
+        public void AddAllowedValues(ITokenString possibleValues)
         {
-            foreach (string value in possibleValues)
-                this.AllowedRootValues.Add(value);
-            this.CanHaveMultipleValuesForRoot = this.CanHaveMultipleValuesForRoot || possibleValues.Count > 1;
+            // can't reasonably generate known values if tokens are involved
+            if (possibleValues.IsMutable || this.AllowedRootValues == null)
+            {
+                this.AllowedRootValues = null;
+                this.CanHaveMultipleValuesForRoot = true;
+                return;
+            }
+
+            // get possible values from literal token
+            InvariantHashSet splitValues = this.ExtractValues(possibleValues);
+            foreach (string value in splitValues)
+                this.AllowedRootValues.Add(value.Trim());
+            this.CanHaveMultipleValuesForRoot = this.CanHaveMultipleValuesForRoot || splitValues.Count > 1;
         }
 
         /// <summary>Set the current values.</summary>
         /// <param name="values">The values to set.</param>
-        public void SetValue(InvariantHashSet values)
+        public void SetValue(ITokenString values)
         {
-            this.Values = values;
+            this.Values = this.ExtractValues(values);
         }
 
         /// <summary>Set whether the token is valid for the current context.</summary>
@@ -72,6 +83,26 @@ namespace ContentPatcher.Framework.Tokens.ValueProviders
             if (input.IsMeaningful())
                 return new[] { this.Values.Contains(input.Value).ToString() };
             return this.Values;
+        }
+
+
+        /*********
+        ** Private methods
+        *********/
+        /// <summary>Get the comma-separated values from a token string.</summary>
+        /// <param name="values">The token string to parse.</param>
+        private InvariantHashSet ExtractValues(ITokenString values)
+        {
+            if (string.IsNullOrWhiteSpace(values?.Value))
+                return new InvariantHashSet();
+            if (!values.IsReady)
+                throw new InvalidOperationException($"Can't get values from a non-ready token string (raw value: {values.Raw}).");
+
+            return new InvariantHashSet(
+                values.Value
+                    .Split(new[] { ',' }, StringSplitOptions.RemoveEmptyEntries)
+                    .Select(p => p.Trim())
+            );
         }
     }
 }